\name{genoConvMat}
\alias{genoConvMat}
\title{
Matrices for Conversion between Multiallelic Genotype Probabilities and
Allele Copy Number Probabilities
}
\description{
By default, this function will produce a matrix that can be used for converting
a matrix of multiallelic genotype probabilities to a matrix of allele copy
number probabilties via matrix multiplication.  If \code{inverse = TRUE}, the
Moore-Penrose pseudoinverse of the matrix is returned, to be used for converting
allele copy number probabilities to approximate multiallelic genotype
probabilities via matrix multiplication.
}
\usage{
genoConvMat(ploidy, n_alleles, inverse = FALSE)
}
\arguments{
  \item{ploidy}{
An integer indicating the ploidy.
}
  \item{n_alleles}{
An integer indicating the number of alleles.
}
  \item{inverse}{
A logical value.  If \code{TRUE}, the pseudoinverse matrix is returned.
}
}
\details{
If we know probabilities of multiallelic genotypes, it is easy to derive
allele copy number probabilities from them.  For example, in a tetraploid
with three alleles, the probabilties of having two copies of allele 0 is
the sum of the probabilities of genotypes 0011, 0012, and 0022.  The
matrix generated when \code{inverse = FALSE} is used for calculating these
sums via matrix multiplication.

\deqn{CM = A}

where \eqn{M} is the matrix of multiallelic
genotype probabilities, with genotypes in rows and individuals in columns,
\eqn{A} is the matrix of allele copy number probabilties, with allele
copy numbers in rows and individuals in columns, and \eqn{C} is the
matrix generated by \code{genoConvMat}.

If instead we know probabilities of allele copy numbers, we can estimate
multiallelic genotype probabilities using the inverse of \eqn{C} (setting the
function to \code{inverse = TRUE}):

\deqn{M = C^{-1}A}

Given that \eqn{C^{-1}} may not be solvable, the pseudoinverse is used, which
represents the shortest length least squares solution to the equation.
}
\value{
If \code{inverse = FALSE}, a matrix with allele copy numbers in rows and 
multiallelic genotypes in columns.  All values are either \code{0} or
\code{1}, indicating whether or not that multiallelic genotype corresponds to
that copy number for that allele.

If \code{inverse = TRUE}, a matrix with multiallelic genotypes in rows and
allele copy numbers in columns, with numbers ranging from -1 to +1.  These
numbers indicate how much each allele copy number probability contributes
to each multiallelic genotype probability.

Rows and columns are named to assist the user with interpretation of
the matrix.  Genotypes are ordered according to the VCF specification.
Alleles are numbered starting at zero.
Allele copy numbers are ordered from \code{0} to \code{ploidy} for allele 
\code{0}, then \code{0} to \code{ploidy} for allele \code{1}, etc.
}

\author{
Lindsay V. Clark
}

\seealso{
\code{\link{enumerateGenotypes}}
}
\examples{
# say we have a tetraploid with three alleles
convmat1 <- genoConvMat(4, 3, inverse = FALSE)
convmat2 <- genoConvMat(4, 3, inverse = TRUE)

# generate some multiallelic genotype probs for this example
genoprobs <- matrix(nrow = 15, ncol = 2,
                    dimnames = list(genotypeStrings(4, 3, sep = ""),
                                    c("ind1", "ind2")))
genoprobs[,1] <- sample(c(5000, sample(100, 14)))
genoprobs[,1] <- genoprobs[,1]/sum(genoprobs[,1])
genoprobs[,2] <- sample(c(900, sample(100, 14)))
genoprobs[,2] <- genoprobs[,2]/sum(genoprobs[,2])

# convert to allele dosage probabilities
alprobs <- convmat1 \%*\% genoprobs

# convert back to multiallelic genotype probabilities
genoprobs2 <- convmat2 \%*\% alprobs
}

\keyword{ algebra }
\keyword{ array }
