\name{markValidity}
\alias{markValidity}
\alias{markValidity,VCF-method}
\alias{validPloidyverseVCF_Precall}
\alias{validPloidyverseVCF_Precall,VCF-method}
\alias{validPloidyverseVCF_Postcall}
\alias{validPloidyverseVCF_Postcall,VCF-method}
\alias{validPloidyverseVCF_Archival}
\alias{validPloidyverseVCF_Archival,VCF-method}
\title{
Check Validity of VCF for Ploidyverse Applications
}
\description{
\code{markValidity} checks a \code{\link[=VCF-class]{VCF}} object to determine
whether it if properly formatted for genotype calling using ploidyverse 
software, whether gentoypes have already been called using ploidyverse 
software, and whether the VCF meets ploidyverse standards for data archiving.
A section is added to the VCF header to indicate the results of this check.

\code{validPloidyverseVCF_Precall}, \code{validPloidyverseVCF_Postcall}, and
\code{validPloidyverseVCF_Archival} examine the header information generated by
\code{markValidity} to return \code{TRUE} or \code{FALSE}.

Together, these functions can be used for error checking before calling 
genotypes or performing downstream analysis.  \code{markValidity} also provides
useful, human-readable information when the \code{VCF} object is written back
to a file.
}
\usage{
markValidity(object)

validPloidyverseVCF_Precall(object)
validPloidyverseVCF_Postcall(object)
validPloidyverseVCF_Archival(object)
}
\arguments{
  \item{object}{
A \code{VCF} object.
}
}
\details{
For a VCF to be valid for "Precall", "Postcall", or "Archival", it must contain
the "AD" genotype field indicating allelic read depth.

For a VCF to be valid for "Postcall", it must contain the "GP" genotype field
indicating genotype posterior probabilities.

For a VCF to be valid for "Archival", it must contain tag sequences in the INFO
column if a non-reference pipeline was used, or contig lengths and assembly 
names if a reference pipeline was used.  It must also contain a "SAMPLE" header
line indicating the species and ploidy of every sample in the file.
}
\value{
\code{markValidity} returns a \code{VCF} object, with a \code{\link{DataFrame}}
called \code{ploidyverseValidity} added to \code{meta(header(object))}.  
There are three rows, named \code{"ploidyversePrecall"},
\code{"ploidyversePostcall"}, and \code{"ploidyverseArchival"}.  A column 
named \code{"Valid"} contains \code{1} if the \code{VCF} object passed the
corresponding checks, and \code{0} if it didn't.  A column named
\code{"Description"} contains human-readable descriptions of the results
of each check.

\code{validPloidyverseVCF_Precall}, \code{validPloidyverseVCF_Postcall}, and
\code{validPloidyverseVCF_Archival} return \code{TRUE} if the corresponding
value in the \code{"Valid"} column is 1, and \code{FALSE} otherwise.
}
\author{
Lindsay V. Clark
}
\seealso{
\code{sampleinfo}
}

\examples{
# To be added
}
\keyword{ methods }
